pyproject.toml
[tool.poetry]
name = "demo-target-2021-05"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2 = "^2.8.6"
Flask = "^2.0.1"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
